Add "agency" table to what-are-agents.mdx
This adds the "agency" table developed for smolagents.
Ethics & society regulars further adapted this table as well, in case you want to use that one (with credit to the fantastic original) -- see here: https://huggingface.co/blog/ethics-soc-7#overview
We:
- Removed "multi-agent" and instead say that these different levels can be combined in multi-agent systems.
- Added "fully autonomous"

Note that we're using the term "autonomy" instead of "agency". We explain more of why in our paper: https://arxiv.org/abs/2502.02649 but the tl;dr is that "agency" is by definition a human term, "autonomy" is more in line with what it means when a computer does it.